% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-get.R
\name{get_model_env}
\alias{get_model_env}
\alias{get_from_env}
\alias{set_env_val}
\title{Working with the modelenv model environment}
\usage{
get_model_env()

get_from_env(items)

set_env_val(name, value)
}
\arguments{
\item{items}{A character string of objects in the model environment.}

\item{name}{A single character value for a new symbol in the model
environment.}

\item{value}{A single value for a new value in the model environment.}
}
\description{
These functions read and write to the environment where the package stores
information about model specifications.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Access the model data:
current_code <- get_model_env()
ls(envir = current_code)

get_from_env("models")
get_from_env("modes")

get_from_env("example")
set_env_val("example", 4)
get_from_env("example")
\dontshow{\}) # examplesIf}
}
