% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_fit.R
\name{is_unsupervised_fit}
\alias{is_unsupervised_fit}
\title{Determine object is has class unsupervised_fit}
\usage{
is_unsupervised_fit(x)
}
\arguments{
\item{x}{an model fit object}
}
\value{
A single logical indicating input has class \code{"unsupervised_fit"}
}
\description{
Determine object is has class unsupervised_fit
}
\examples{
model_fit <- list(letters)

is_unsupervised_fit(model_fit)

model_fit <- new_unsupervised_fit(model_fit)

is_unsupervised_fit(model_fit)
}
\seealso{
\code{\link[=new_unsupervised_fit]{new_unsupervised_fit()}}
}
