% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_spec.R
\name{is_unsupervised_spec}
\alias{is_unsupervised_spec}
\title{Determine object is has class unsupervised_spec}
\usage{
is_unsupervised_spec(x)
}
\arguments{
\item{x}{an model specification}
}
\value{
A single logical indicating input has class \code{"unsupervised_spec"}
}
\description{
Determine object is has class unsupervised_spec
}
\examples{
model_spec <- list(letters)

is_unsupervised_spec(model_spec)

model_spec <- new_unsupervised_spec(model_spec)

is_unsupervised_spec(model_spec)
}
\seealso{
\code{\link[=new_unsupervised_spec]{new_unsupervised_spec()}}
}
