% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelenv-package.R
\docType{package}
\name{modelenv-package}
\alias{modelenv}
\alias{modelenv-package}
\title{modelenv: Provide Tools to Register Models for Use in 'tidymodels'}
\description{
An developer focused, low dependency package in 'tidymodels' that provides functions to register how models are to be used. Functions to register models are complimented with accessor functions to retrieve registered model information to aid in model fitting and error handling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/modelenv}
  \item \url{http://modelenv.tidymodels.org/}
  \item Report bugs at \url{https://github.com/tidymodels/modelenv/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
