% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_fit.R
\name{new_unsupervised_fit}
\alias{new_unsupervised_fit}
\title{Give object unsupervised fit class}
\usage{
new_unsupervised_fit(x)
}
\arguments{
\item{x}{an model fit object}
}
\value{
x with added class \code{"unsupervised_fit"}
}
\description{
Give object unsupervised fit class
}
\examples{
model_fit <- list(letters)
model_fit <- new_unsupervised_fit(model_fit)
model_fit
}
\seealso{
\code{\link[=is_unsupervised_fit]{is_unsupervised_fit()}}
}
