% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumervised_spec.R
\name{new_unsupervised_spec}
\alias{new_unsupervised_spec}
\title{Give object unsupervised specification class}
\usage{
new_unsupervised_spec(x)
}
\arguments{
\item{x}{an model specification}
}
\value{
x with added class \code{"unsupervised_spec"}
}
\description{
Give object unsupervised specification class
}
\examples{
model_spec <- list(letters)
model_spec <- new_unsupervised_spec(model_spec)
model_spec
}
\seealso{
\code{\link[=is_unsupervised_spec]{is_unsupervised_spec()}}
}
