% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_fit.R
\name{stop_incompatible_mode}
\alias{stop_incompatible_mode}
\title{Error handling for incompatible modes}
\usage{
stop_incompatible_mode(spec_modes, eng = NULL, model = NULL)
}
\arguments{
\item{spec_modes}{Character vector of modes}

\item{eng}{Character of specific engine}

\item{model}{Character of specific model}
}
\value{
An error
}
\description{
Error handling for incompatible modes
}
\examples{
library(rlang)
tmp <- catch_cnd(stop_incompatible_mode("partition"))
}
