% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_spec_mode_engine_val}
\alias{check_spec_mode_engine_val}
\title{Error handling for unknown mode}
\usage{
check_spec_mode_engine_val(model, mode, eng)
}
\arguments{
\item{model}{Character of specific model}

\item{mode}{Character of specific mode}

\item{eng}{Character of specific engine}
}
\value{
An error
}
\description{
Checks that a given model, mode, engine combination have been registered.
}
\examples{
library(rlang)
tmp <- catch_cnd(check_spec_mode_engine_val("turtle", "partition", "vegan"))
}
